Log improvements - WIP: - remove useless code - fix namespace (remove all log sub-folder because we are in log plugin)
