+++ src/main/core/Manager/Workspace/WorkspaceManager.php
@@ -24,12 +24,12 @@ use Claroline\CoreBundle\Entity\Workspace\Shortcuts;
 use Claroline\CoreBundle\Entity\Workspace\Workspace;
 use Claroline\CoreBundle\Entity\Workspace\WorkspaceOptions;
 use Claroline\CoreBundle\Entity\Workspace\WorkspaceRegistrationQueue;
-use Claroline\LogBundle\Event\CatalogEvents\SecurityEvents;
-use Claroline\LogBundle\Event\Security\AddRoleEvent;
 use Claroline\CoreBundle\Manager\ResourceManager;
 use Claroline\CoreBundle\Manager\RoleManager;
 use Claroline\CoreBundle\Repository\User\UserRepository;
 use Claroline\CoreBundle\Repository\WorkspaceRepository;
+use Claroline\LogBundle\Event\CatalogEvents\SecurityEvents;
+use Claroline\LogBundle\Event\Security\AddRoleEvent;
 use Doctrine\Persistence\ObjectRepository;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerInterface;
