+++ src/main/log/API/Serializer/FunctionalLogSerializer.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Claroline\LogBundle\API\Serializer\Log;
+namespace Claroline\LogBundle\API\Serializer;
 
 use Claroline\AppBundle\API\Options;
 use Claroline\AppBundle\API\Serializer\SerializerTrait;
@@ -8,7 +8,7 @@ use Claroline\CoreBundle\API\Serializer\Resource\ResourceNodeSerializer;
 use Claroline\CoreBundle\API\Serializer\User\UserSerializer;
 use Claroline\CoreBundle\API\Serializer\Workspace\WorkspaceSerializer;
 use Claroline\CoreBundle\Library\Normalizer\DateNormalizer;
-use Claroline\LogBundle\Entity\Log\FunctionalLog;
+use Claroline\LogBundle\Entity\FunctionalLog;
 
 class FunctionalLogSerializer
 {
