+++ src/main/log/Resources/config/services/finder.yml
@@ -1,6 +1,6 @@
 services:
-    Claroline\LogBundle\API\Finder\Log\SecurityLogFinder: ~
+    Claroline\LogBundle\API\Finder\SecurityLogFinder: ~
 
-    Claroline\LogBundle\API\Finder\Log\MessageLogFinder: ~
+    Claroline\LogBundle\API\Finder\MessageLogFinder: ~
 
-    Claroline\LogBundle\API\Finder\Log\FunctionalLogFinder: ~
+    Claroline\LogBundle\API\Finder\FunctionalLogFinder: ~
